Add crawl.get_target
This clua function opens the targeting dialog in the same way it is
opened when you press x to examine. The user is allowed to move the
cursor around (same as examine). If the user makes a selection and
presses enter, this function returns the player-relative coordinates of
the selection. Note that the target is valid even if it is past the
bounds of the map, so there is no leak of world borders. If the user
cancels the selection, this function returns nil.