Handle hidpi screens in webtiles glyph mode
This was not factoring in the device pixel ratio correctly, and so ended
up too small on hidpi screens. This commit also adds support for
crawl-internal zooming (via rc options, or {} in X) of the dungeon.
Browser zoom affecting glyph mode is a different story, I'm not sure how
possible this is going to be.