Show XL to reach relevant skill level on more items (Mantis #12018)
Before, if you could not *currently* train a particular skill, you would
not see the "At X% training you would reach Y in about Z XLs" line. For
example, if you did not have any short blades in your inventory but were
looking at a rapier in a shop, the line would be missing. This fixes the
description code to only exclude the line if a skill is *permanently*
untrainable.