Don't pull abyss terrain from the vestibule
The issue is that if the player hasn't visited the vestibule, a level
excursion there will trigger asserts regarding the up stairs, which
haven't been set up yet. This could be made conditional on having
visited the vestibule, but even for the abyss I'd rather avoid
dependencies on level traversal order.