Clean up some heavy-handed artefact code
Any time you checked an artefact property, the old code made an entire
copy of the CrawlHashTables for both the artefact props and its known
state. Because artefact properties are checked frequently when monsters
are carrying them (mostly in case they have flight), this could lead to
heavy world_reacts on some levels, e.g. Vaults 5. This was particularly
noticeable with step from time on a debug build.