Simplify some holiness checks for monster swapping
One of these checks changes behavior. Requiring the same holiness for
swapping around intelligence in general seems really spoilery and weird.
If this doesn't work out (e.g. there's too much swapping as a result, or
some noticeable effect on balance) I'm open to reverting or modifying
this. However, I think if something like this needs to be in place, then
we should first try a more obvious rule, e.g. nothing can swap past
undead.  (Which I think is the main use case anyways.)

The demonic change is just an optimization and shouldn't affect
behavior. (Full holiness checks are costly if done a lot, because they
check damage types -- but this is irrelevant for the demonic check.)