Handle HUPs better in lua interpreter
Previously, this code ran in a tight loop if a hup was received, because
the input function always returned immediately. This led to stuck
high-cpu processes and crashes when a webtiles client disconnected, or
when an ssh connection was lost, while the lua interpreter was running.
This commit should simply break the loop on this case. It's possible it
would be even safer to not do a repl loop at all on empty input for
online games...