Add a placeholder fix for trap+corruption+nymphs
What seems to be happening is that a trap gets temporarily removed by a
nymph's flooding, then the terrain change marker gets moved by
corruption without moving the corresponding env.trap entry, then the
game crashes when restoring. This fix, which is not intended to be
final, will instead print an error and place floor on the trap reversion.

I think a real fix should either check and move the trap_def when the
terrain change marker moves, or remove and recreate env.trap entries in
sync with any terrain changes. The latter seems better, but I'm not sure
why it wasn't implemented that way in the beginning, so there may be
some issue I'm not anticipating.