De-closet the central area of an encompass vault
It's possible for a formicid to get stuck in the central safe area in
hangedman_spin_cycle, since this vault can place on branch ends, which
wouldn't allow shafting out. The area can't have no_tele_into otherwise
it's totally inaccessible outside of wall breaking. The only
alternatives to fix it are to fill it with wall or to make the area
universally accessible. I went with the latter approach, since it's more
fun and flavorful.

This commit adds stone glass so that most of the area can be seen into
and randomly places two transporters to provide entry and exit. Anyone
shafting into the central area can use the exit transporter, and players
from outside can enter the safe area knowing what they'll encounter, for
the most part. Note that it's still possible for uniques like the
Enchantress to place in the central area or for monsters to just
teleport or shaft in, so it's not always fully safe!

I left in the 'exit' teleport traps since that possibly provides a
better exit method compared to the exit transporter, which could have
monsters around it. The very center and some outer edges of the central
area can't be seen through the glass, something which could be cleaned
up in the future.

The original vault author can feel free to try another approach to this
fix so long as shaft closet issue is fixed somehow.