Fill small boring disconnected areas in the builder
Often, when vault placement doesn't happen to align well with the level
layout, small bits of the level might get cut off and isolated. This
will trigger a veto, making hard-to-place vaults even rarer. This is a
reasonably common source of vetoes in depths, in particular, in my
testing. This commit adds a post-place heuristic that attempts to fill
in small disconnected regions with no stairs.

This is effective in many cases, but while vaults default to
non-transparent, it is risky: the builder relies on an opaque vault's
exit being connected to something that is not masked as opaque, and uses
that something as a proxy for level connectivity in general. So if an
opaque vault is connected to a small space that this change would fill
in, that vault gets disconnected from the rest of the world. A future
commit will deal with this issue by making exits transparent even on
opaque vaults.