Don't by default mask vault exits as opaque
The problem this commit is solving is that opaque vaults rely on their
exits being connected to some non-vault squares in order to properly
test for connectivity, and I want to apply post-placement heuristics
that do things like fill small disconnected areas (as in preceding
commit). This change allows an opaque vault to veto correctly if its
exit ends up leading to a wall because of such heuristics. It would be
possible to explicitly mask exits as opaque in order to add optional
exits; this commit effectively forces all exits to participate in
connectivity by default.

For a completely internally-connected vault with multiple exits, this
commit may make placement harder. However, the solution for such vaults
is simply to mark that vault as transparent. (Possibly, all vaults
should be transparent by default.)