Fix a potential crash while joining ru
This is a quick fix that I'm not sure is ideal, but what seems to be
happening is that your_talents can be triggered as a side effect of
messaging while joining a god, which then trips a Ru ASSERT before the
god is fully set up. Changing the order of calls in join_religion might
be better, but that sequence looks very delicate.