Copy .md files on install
I think that before bf3a5c7, this is what was causing various crashes in
circumstances that aren't what we usually test. This commit isn't
strictly necessary after bf3a5c7, but it paves the way for converting
other docs to md, and also heads off this gotcha for the future.

Resolves #1609