Fix a console hud issue at high turncounts (gressup)
This code used a constant number of spaces to clear the rest of the
turncount line, but didn't factor in the fact that cprintf wraps; at
high turncounts (6+ digits) on long autoexplore/autotravel spans (>10.0
auts) it could therefore overwrite 1-2 letters of the weapon display.
Recent changes reduce how often the weapon redraws, so this has probably
become more visible lately. Despite that, this is a vintage bug,
originating from ce229f1793ccd.