Limit printable keycodes in webtiles to ASCII
This was (I think) impacting non-US keyboards, where keycodes
corresponding to invalid UTF-8 were causing the webserver to choke, and
the macro menu not to get forwarded to the client (keycode 169, which
*may* result from a closing paren on AZERTY, is one example). It
possibly also affects some unusual keys on certain keyboards. There must
be a more elegant solution but for now this will just show the keycode
directly in the macro menu. Better would be to somehow figure what keys
are involved, and print their names, but this seems a bit of a hopeless
task for webtiles.

It's possible that console needs a similar fix as well, but there (at
least in my testing) the result is not that the menu fails altogether --
though it's not very usable. But this would be better tested by someone
with a different locale / keyboard layout. SDL should be handled by SDL
code, so if it fails on this stuff, it's at least not a crawl bug.