merge [Add a basic menu to macro/keybind editing]
Especially for online play, it's extremely annoying that there's no way
to view current macro/keymap mappings without going key by key. This
commit solves that by putting a real submenu behind the `~` options,
letting you see all current macros/keybinds (of the same type) at once.
It also does some light refactoring of how some of the options are
chosen, and will involve an extra keypress for new macros (or more for
keys that are used in the menu ui).  This menu is very much at the
minimal viable ui state: I didn't widgetify anything except the
key/macro selection part, and there's a *lot* more one could do with
this to give it a real ui. However, for now it's a net improvement.

# Conflicts:
#	crawl-ref/source/macro.cc
#	crawl-ref/source/webserver/game_data/static/menu.js