Convert traversability memoization to an array
a42c0fa56e7 did get noticeable speedups in world_reacts bad cases, but
it turns out that lookup time for unordered_map on this scale is still
surprisingly bad because (it seems) of weird stdlib implementation
constraints that lead to high layers of abstraction surrounding the
constant lookup. So, just use an array, like most other member functions
for this class.