Dwarven Slime layout groundwork: tweak parameters and remove glass
For now we just tweak the parameters of layout_geoelf_grid and
make glass not generate in Slime. A small amount of glass is
removed from Elf for convenience (testing suggested the difference
was minor).

Currently, fancy room fraction is set to 0. A subsequent commit
will make fancy rooms with Dwarven features.