partial merge [New Slime monster: Rockslime]
Rockslimes are slimes that trample you, perhaps into
the nearest slime wall. They have high AC, rF++ and rElec,
since they're made of rock, and wallop for a hefty 50 per
hit. They're intended to be on par with azure jellies and
a bit lower-threat than acid blobs (they have lower HP than
either), and to play with the Slime terrain in a fun way.

# Conflicts:
#	crawl-ref/source/monster-type.h
#	crawl-ref/source/spl-damage.cc