Have soul bound monsters always create simulacra
The monster Bind Souls spell currently only creates simulacra if the
monster rolled successfully for the chance to leave a corpse. This is an
aspect of how the old version of the spell directly consumed corpses,
which closely mirrored the player Simulacrum spell. However Bind Souls
isn't trying to have symmetry with the player spell; in fact the two
spells don't have same name. The corpse requirement only serves to
weaken the most distinctive ability of the necromancer monster, which
isn't very dangerous at the depths in which it appears.

This commit has a simulacra always be created when a monster dies with
the Bind Souls enchantment active, giving Bind Souls and by extension
necromancers a buff.