fix: don't overwrite (jobbed) draconian breath spells
A draconian with a job usually has a spellbook, and this code would
overwrite their breath weapon with that spellbook. This appears to be
unintentional, and is pretty old. The bug was not apparent because the
describe code doesn't actually use the monster spell list to show the
breath weapon; see mon-util.cc `get_unique_spells` and
describe-spells.cc `_monster_spellbooks`. (In general, this code is not
what I expected and is probably ripe for bugs like this...)