fix: expand opaque mask skipping for vault exits
Since 080276f, the `@` glyph is never automatically masked as opaque,
leading to some connectivity checks for non-transparent vaults that do
have exits. However, many vaults with only one exit rely on `+`, `=`, or
even ` ` (in the case of floats with no marked exits) being interpreted
as an exit as well. Example: library_closet in
https://crawl.develz.org/mantis/view.php?id=12591. This commit expands
things so that anything that counts as an exit for a vault will not
receive an opaque mask unless that glyph is explicitly masked, which
will expand connectivity checks by quite a bit. This may over-veto,
we'll see, but in my testing seems to work and fixes the issue in 12591
without otherwise disrupting the seed (will need to do some more
automated testing).