fix: prevent closets in `heliophobic_arrival_ceremonial_halls`
Extends the fix in #1984. The biggest thing this vault needed was a
`transparent` tag; I've also added the pathing from #1981 as it will
help it place, and in addition to that added explicit @ tags. It could
be good to randomize the outer pathing a bit, but let's keep it simple
for now.

Resolves #1973
Resolves #1984