Adjust spawns in a vault (dilly)
The vault nicolae_shoals_spring_well_temple places about 7.5 water
nymphs on average, in addition to two well springs, which is pretty
extreme. The upper-limit of what vaults can place given extreme and rare
RNG results is allowed to be ridiculous, sometimes worse than branch
endings and potentially leading the player to abandon a level. However
unless the vault is intended to be an extreme challenge, which often
means down-weighting or runed door/transporter vault, the average case
should be more reasonable.

This commit adjusts the vault to only place 3 water nymphs on-average
and adjust the well spring placement from guaranteed 2 to place 1 to 3
with an average of two.