feat: Some tweaks for a Gauntlet exit map
For dreamdust_gauntlet_exit_labyrinth, use alternating patterns to place
dead ends not cardinally adjacent to wall openings, which would make the
dead end not function. Tweak the inner-most wall to always a path around
corners that doesn't create a diagonal choke point. Finally, move the
map to the end of the DES section, so that the maps are ordered by when
they were added. Closes #2166.