New unique: Zenata, Seeker of Sheza
Zenata is another late-game unique (appearing in Depths, Zot &c),
borrowing a bit from the Desolation of Salt and putting a twist on it.
She's taken the ragged hierophants' "Resonance Strike", which is like
airstrike that gets bonus damage from adjacent constructs instead of
from open spaces, and uses it in conjuction with a new "Sheza's Dance"
spell, which summons 2-4 dancing weapons. To encourage players to kill
her, she comes with a high-tier weapon with some plusses already on it.

Zenata's theme is intended to expand slightly on the Desolation lore.
She entered the Dungeon to look for secrets of the old Sheza/Sargol cult,
the Divine Twins that ran the Desolation of Salt before it was a
desolation. The bits and pieces she found required her to replace more
and more of her flesh with metal to unlock their power... but hey,
upgrade! :)

# Conflicts:
#	crawl-ref/source/dat/des/builder/uniques.des
#	crawl-ref/source/mon-cast.cc
#	crawl-ref/source/mon-gear.cc
#	crawl-ref/source/monster-type.h
#	crawl-ref/source/spell-type.h
#	crawl-ref/source/spl-data.h